move code all into one file in FerrostarCore. Now @retroactive is not necessary? This is good, but I'm confused why I'd needed it before! :)
