+++ .gitattributes
@@ -4,4 +4,4 @@ android/demo-app/** -linguist-detectable
 
 apple/DemoApp/** -linguist-detectable
 
-apple/Sources/UniFFI/ferrostar.swift linguist-generated=true
+apple/Sources/UniFFI/** linguist-generated=true
